Remove deny_info clause, deny blacklist only for logged users
Deny_info is not working for HTTPS connections. It's impossible to
redirect and HTTPS CONNECT request. The only feasible way to do it
is to "man-in-the-middle" the connection: perform an SSL handshake,
decrypt the request, return an error page through the
browser-proxy HTTPS connection. This involves a privacy break for
any other connection to the allowed websites.