Change from redirector to external_acl, with custom error pages redirections
Now the db_blacklist_helper.py (still to fix and adapt) will work as an external
acl instead of a redirector program. The helper will recieve the url, username,
referer and (big news) the mime type OF THE HTTP REPLY. With this data it will
be able to decide if the requested URL is just a page resource and allow it or
not.

When the URL is not allowed, the user is 307-redirected (thanks to the deny_info
directive) to a custom error page (still to do). The URL of this error page
contains a querystring with the username, requested url and error message
from the helper script. With this data it will be easy to generate a dynamic
form for limbo motivations, for unlock requests or ban notices.